+++ src/query-hooks/useAuth/POST/index.ts
@@ -1,0 +1,1 @@
+export { default } from './usePostAuth';
