+++ src/query-hooks/useAuth/index.ts
@@ -1,0 +1,5 @@
+import POST from './POST';
+
+export default {
+    POST
+};
