+++ src/query-hooks/useUser/GET/index.ts
@@ -1,1 +1,1 @@
-export { default } from './useFetchLoginStatus';
+export { default } from './useGetUser';
