+++ src/query-hooks/useUser/POST/apis.ts
@@ -1,8 +1,7 @@
 import axiosInstance from '@/lib/axiosInstance';
 import sha256 from 'crypto-js/sha256';
-import Cookies from 'js-cookie';
 
-import type { CreateNewUser, PostAuthenticateUser } from './apis.type';
+import type { CreateNewUser } from './apis.type';
 
 export const createNewUser: CreateNewUser.Func = async ({
     nickname,
@@ -24,31 +23,6 @@ export const createNewUser: CreateNewUser.Func = async ({
         return data;
     } catch (err) {
         console.error('유저 생성을 실패했습니다.', err);
-        throw err;
-    }
-};
-
-export const postAuthenticateUser: PostAuthenticateUser.Func = async ({
-    email,
-    password
-}) => {
-    try {
-        const params = <PostAuthenticateUser.ReqParams>{
-            email,
-            password: sha256(password).toString()
-        };
-
-        const { data } =
-            await axiosInstance.post<PostAuthenticateUser.Response>(
-                'auth/login/local',
-                params
-            );
-
-        Cookies.set('accessToken', data.accessToken);
-
-        return data;
-    } catch (err) {
-        console.error('로그인을 실패했습니다.', err);
         throw err;
     }
 };
